backend/flask/requirements.txt
apscheduler
gunicorn
Flask
flask-cors
flask-restful
onesignal-sdk
pymysql
python-dotenv
requests
sqlalchemy